[UI] Добавлен раздел "Друзья" пользователя
Resolve #626
