+++ src/frontend/components/user/status.tsx
@@ -1,0 +1,54 @@
+"use client";
+import { Button } from "@heroui/button";
+import toast from "react-hot-toast";
+import { Chip } from "@heroui/chip";
+
+import { FriendStatus } from "@/lib/models";
+import { AddFriend } from "@/lib/requests";
+
+const UserStatus = ({
+  friendStatus,
+  userId,
+  setFriendStatus,
+}: {
+  friendStatus: FriendStatus;
+  userId: number;
+  setFriendStatus: { (status: FriendStatus): void };
+}) => {
+  let item;
+  switch (friendStatus) {
+    case FriendStatus.no_friend:
+      item = (
+        <Button
+          variant="ghost"
+          onPress={async () => {
+            const result = await AddFriend(userId);
+            if (result && "code" in result) {
+              toast.error(result.message);
+            } else {
+              toast.success("Заявка отправлена");
+              setFriendStatus(FriendStatus.has_outcoming_request);
+            }
+          }}
+        >
+          Добавить в друзья
+        </Button>
+      );
+      break;
+    case FriendStatus.has_incoming_request:
+      item = <Chip color="warning">Отправил вам заявку в друзья</Chip>;
+      break;
+    case FriendStatus.has_outcoming_request:
+      item = <Chip color="warning">Заявка отправлена</Chip>;
+      break;
+    case FriendStatus.is_friend:
+      item = <Chip color="success">Ваш друг</Chip>;
+      break;
+    case FriendStatus.is_yourself:
+      item = <Chip color="success">Вы</Chip>;
+  }
+
+  return item;
+};
+
+export default UserStatus;
